Added fonts and basic navbar minimal styling
