+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,21 @@
+.navbar{
+    display: flex;
+    flex-direction: row;
+    color: white;
+}
+ul {
+    display: flex;
+    flex-direction: row;
+    margin-left: auto;
+    margin-right: auto;
+    align-items: stretch;
+    justify-content: space-around;
+
+}
+li {
+    margin: 1vw;
+    justify-content: space-around;
+}
+.title {
+    color: #FFC904;
+}
