+++ src/components/Navbar/index.jsx
@@ -1,0 +1,15 @@
+import logo from '../../assets/logo.png'
+import './Navbar.css'
+function Navbar(){
+		return (
+		<nav className='navbar'>
+			<h1 className='title'>Knightro Run</h1>
+			<ul> 
+				<li><a href="#">Sign up</a> </li> 
+				<li> <a href="#">Sign in</a> </li>
+			</ul>
+		</nav>
+		);
+	
+}
+export default Navbar
